% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybuilder.R
\name{cox_KM_plots}
\alias{cox_KM_plots}
\title{Generate Cox predicted KM plots}
\usage{
cox_KM_plots(KM.hat, clrs = NULL)
}
\arguments{
\item{KM.hat}{Kaplan-Meier estimate}

\item{clrs}{color of lines, consider for removal}
}
\value{
Plot of predicted survival curves
}
\description{
Generate Cox predicted KM plots
}
