% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybuilder.R
\name{cox_times_table}
\alias{cox_times_table}
\title{Create table of Cox predicted probabilities}
\usage{
cox_times_table(KM.hat, fixTimes = NULL)
}
\arguments{
\item{KM.hat}{Survival estimator}

\item{fixTimes}{vector of times}
}
\value{
Table of predicted probabilities for each time given
}
\description{
Create table of Cox predicted probabilities
}
