% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_blocks.R
\name{ez_pickone}
\alias{ez_pickone}
\title{Generate code for \code{selectInput()}}
\usage{
ez_pickone(vname, label, choices)
}
\arguments{
\item{vname}{variable name, or inputId}

\item{label}{visual label of select}

\item{choices}{vector of choices to be made available}
}
\value{
Vector containing ui and server code
}
\description{
Creates a \code{selectInput()} function in shiny ui and pairs it with a server
input.
}
