% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_blocks.R
\name{ez_pickone_logic}
\alias{ez_pickone_logic}
\title{Generate code for \code{selectInput()}}
\usage{
ez_pickone_logic(vname, label, choices)
}
\arguments{
\item{vname}{variable name, or inputId}

\item{label}{visual label of select}

\item{choices}{vector of choices, must be TRUE and FALSE}
}
\value{
vector containing ui and server code for \code{selectInput()}
}
\description{
Creates a \code{selectInput()} function in shiny ui and
pairs it with a server input. Only for logical variables
}
