% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_blocks.R
\name{ez_slider}
\alias{ez_slider}
\title{Generate code for \code{sliderInput()}}
\usage{
ez_slider(vname, label, min, max, value)
}
\arguments{
\item{vname}{Name of variable}

\item{label}{Label to display in shiny ui}

\item{min}{minimum value of slider}

\item{max}{maximum value of slider}

\item{value}{default slider value}
}
\value{
Vector containing shiny ui and server code for \code{sliderInput()}
}
\description{
Creates shiny ui code for \code{sliderInput()} and creates input for vname in
server
}
