% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confints.R
\name{get_confint}
\alias{get_confint}
\title{Creates confidence levels for plotting predicted survival curves.}
\usage{
get_confint(p, se, logse = TRUE, conf.type, conf.int, selow, ulimit = TRUE)
}
\description{
Creates confidence levels for plotting predicted survival curves.
}
\references{
survival authort
}
