% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confints.R
\name{part2}
\alias{part2}
\title{Creates predicted survival and standard errors for confidence intervals}
\usage{
part2(listsurv, coxfit, newdata)
}
\arguments{
\item{listsurv}{Output from 'surv_pred_info' function}

\item{coxfit}{coxfit object created for predictions. Used to find strata}

\item{newdata}{Data used to make predicted standard errors}
}
\description{
Creates predicted survival and standard errors for confidence intervals
}
