% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_predictions.R
\name{predict_one_coxfit}
\alias{predict_one_coxfit}
\title{Compute predicted survival outcome}
\usage{
predict_one_coxfit(coxfit, newdata)
}
\arguments{
\item{coxfit}{This is an object returned by \code{\link[=prep_coxfit]{prep_coxfit()}}}

\item{newdata}{vector of new data}
}
\value{
data.frame of predicted survival outcomes, one column is time, one
is probability
}
\description{
Computes predicted survival outcome for one new data row using coxfit list
object created by \code{\link[=prep_coxfit]{prep_coxfit()}}.
}
\section{Note}{

This function's primary use is within the shiny app, where a \code{coxph} object
is not available. It can be used outside of that context but that is the
main purpose of this function, and why it only accepts the return object
of \code{\link[=prep_coxfit]{prep_coxfit()}}.
}

