% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_predictions.R
\name{prep_coxfit}
\alias{prep_coxfit}
\title{Create simplified \code{coxph()} object}
\usage{
prep_coxfit(coxph.result, tol = 1e-07)
}
\arguments{
\item{coxph.result}{Result returned by \code{coxph()}}

\item{tol}{numerical tolerance for prediction differences}
}
\value{
Simplified \code{coxph()} result, in list form, include what's in it
}
\description{
Simplifies \code{coxph()} output and checks that predictions
match those of the full model
}
