% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine_coxph.R
\name{shine_coxph}
\alias{shine_coxph}
\title{Build a shiny app for coxph models}
\usage{
shine_coxph(..., app.dir = NULL)
}
\arguments{
\item{...}{Any number of coxph models}

\item{app.dir}{directory where shiny app will be written}
}
\value{
A shiny app written into specified directory.
}
\description{
Generates a shiny app to visualize Cox proportional hazard models created by
\code{coxph()}
}
\section{Details}{


One of the important features of this function is that the shiny app, once
created, will not contain any identifiable data, containing only
information necessary for predictions. There are some requirements in order
for this function to run without error: in your original \code{coxph()} function
or functions, model = TRUE and x = TRUE are required arguments. Because of
this requirement the use of \code{tt()} is not allowed in our function. As well
as that, strata by covariate interaction terms are not allowed. Please do
not use "." in the formula at this time either. There is a function
included in the package that will handle these requirements and provide
warnings if your model as constructed is unlikely to work with this
function.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

  library(survival)
  temp_app_dir <- tempdir()
  names(leukemia)[names(leukemia) == "x"] <- "treatment"
  model1 <- coxph(Surv(time, status) ~ treatment,
  leukemia, x=TRUE, model=TRUE)

  shine_coxph("Model 1" = model1, app.dir =
  temp_app_dir)

  filedir <- list.files(temp_app_dir)[1]

  shiny::runApp(paste0(temp_app_dir, "/", filedir))
  files <- list.files(temp_app_dir)
  file.remove(files)

\dontshow{\}) # examplesIf}
}
