% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_predictions.R
\name{predict_one_coxfit}
\alias{predict_one_coxfit}
\title{Compute predicted survival outcome}
\usage{
predict_one_coxfit(coxfit, newdata)
}
\arguments{
\item{coxfit}{This is an object returned by \code{\link[=prep_coxfit]{prep_coxfit()}}}

\item{newdata}{vector of new data}
}
\value{
data.frame of predicted survival outcomes
}
\description{
Computes predicted survival outcome for one patient based
on a Cox model fit.
}
