% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine_coxph.R
\name{shine_coxph}
\alias{shine_coxph}
\title{Build a shiny app for coxph models}
\usage{
shine_coxph(...)
}
\arguments{
\item{...}{Any number of coxph models}

\item{app.dir}{directory where shiny app will be written}
}
\value{
A shiny app written into specified directory.
}
\description{
Generates a shiny app to visualize Cox proportional hazard models
created by \code{coxph()}
}
\section{Details}{


One of the important features of this function is that the shiny app, once
created, will not contain any of the original data, with predictions made
using the baseline survival from \code{survfit.coxph()}. Some extra things to
consider. For the function to run properly model = TRUE and x = TRUE are both
required in the \code{coxph()} argument, which limits the ability to use \code{tt()}
functions in the model. As well as that, strata by covariate interaction
terms are not allowed. We have developed a wrapper function \code{\link[=make_coxph]{make_coxph()}},
which will have x = TRUE and model = TRUE as defaults, and also provide
warnings if the model you are using is not suitable for this function.
}

