% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine_coxph.R
\name{shine_coxph}
\alias{shine_coxph}
\title{Build a shiny app for Cox proportional hazard models}
\usage{
shine_coxph(..., app.dir = NULL, theme = c("default", "dashboard"))
}
\arguments{
\item{...}{Objects created by \code{coxph()}}

\item{app.dir}{Directory where shiny app will be written}

\item{theme}{Theme of shiny app. This changes the overall appearance of the
app. The first is the default theme, requiring only shiny to use. The
"dashboard" theme requires the \code{shinydashboard} and \code{DT} packages
to run properly.}
}
\value{
A shiny app written into specified directory.
}
\description{
Generates a shiny app to visualize Cox proportional hazard models created by
\code{coxph()}
}
\section{Details}{


One of the important features of this function is that the shiny app, once
created, will not contain any identifiable data, containing only
information necessary for predictions. There are some requirements in order
for this function to run without error: in your original \code{coxph()} function
or functions, model = TRUE and x = TRUE are required arguments. Because of
this requirement the use of \code{tt()} is not allowed in our function. As well
as that, strata by covariate interaction terms are not allowed. Please do
not use "." in the formula at this time either. There is a function
included in the package that will handle these requirements and provide
warnings if your model as constructed is unlikely to work with this
function.

There are some more guidelines for formula notation. There is some leniency
in that you can use functions inside the formula, but using nested
functions will fail. Considering the input names used in shiny are taken
directly from the formula, it is best not to add too many convolutions.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

  library(survival)
  temp_app_dir <- tempdir()
  names(leukemia)[names(leukemia) == "x"] <- "treatment"
  model1 <- coxph(Surv(time, status) ~ treatment,
  leukemia, x=TRUE, model=TRUE)

  shine_coxph("Model 1" = model1, app.dir =
  temp_app_dir)

  filedir <- list.files(temp_app_dir)[1]

  shiny::runApp(paste0(temp_app_dir, "/", filedir))
  files <- list.files(temp_app_dir)
  file.remove(files)

\dontshow{\}) # examplesIf}
}
