% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confints.R
\name{survflag}
\alias{survflag}
\title{For  use in internal \code{\link[=surv_pred_info]{surv_pred_info()}} function, borrowed from \code{survival}
package}
\usage{
survflag(y, id)
}
\arguments{
\item{y}{Surv object}

\item{id}{vector of ids coinciding with y.}
}
\description{
For  use in internal \code{\link[=surv_pred_info]{surv_pred_info()}} function, borrowed from \code{survival}
package
}
